requirements.txt
Django==2.0
django-filter==2.0.0
djangorestframework==3.8.2
wagtail>=2.2,<2.3
django-storages==1.6.6
boto3==1.8.0
psycopg2-binary==2.7.5
djangorestframework-jwt==1.11.0
django-debug-toolbar==1.11
